+++ src/pages/ws/ui/variant-drawer/variant-drawer-header/drawer-layout-control/preset-menu/index.ts
@@ -1,1 +1,1 @@
-export * from './preset-menu'
+export * from './preset-menu-popover'